move per-group-context actor out of ThumbnailCtx
